Added an optimized random decision maker (relying on the ListAllPossibleDecisiosn to be quicker, since itself relies on the disctionnary)
Added a decision maker based on optimization of scores -> in case of equality, possibility to choose the anti-scopa strategy (leave a board higher than 10). and then in case of new equality it is just the first value...

Also: there is a discrepancy between my way of coding the absence of take ("none") and Gui's way (NULL) -> should eventually be fixed by replacing the "none" by NULL in the play_take_dict. For now there are some mini fixed inside functions but this hsould not stay like this